tech: Add installer
This "installer" is simply a way of making the scripts available to
the shell on the system by putting them inside `~/.local/bin/`-- either
through copying (so these scripts are essentially frozen and require a
manual `make install` to update), or through a Symlink (which automatically
updates those scripts provided you keep this directory in the same place).

I recommend a hard copy to avoid breaking the symlink. The link option is there
for developers.